Changing Density in Marcy_VC_comparison
I realized that we had density a bit off in the marcy file.  We need to
get the plot level density I think, (trees in a plot / plot area).But
things get weird.  I think maybe we do this later and this step isn’t
necessary, but take a look at lines 30-50 and let me know what you
think.  So after writing this I think we assign a density to a single
tree, and then add up the trees in a plot to get the density for a
plot.  But I end up confusing myself.  Take a look and see what you
think.